+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # v0.2.2.rc1 (git master)
 
+- ProxyCollection destroy_all and delete_all returns removed records.
 - Changes are better detected on new records. user = User.new :email => "me@gmail.com"; user.email_changed?
   returns true.
 - Mass assignment of attributes now support multi parameter so we should be able to support form helpers